コントローラーの記述を簡略化。

+++ app/controllers/likes_controller.rb
@@ -1,21 +1,24 @@
 class LikesController < ApplicationController
+
+  before_action :set_prototype, before: [:create, :destroy]
+
   def create
     current_user.likes.create(create_params)
-    @prototype = Prototype.find(params[:prototype_id])
   end
 
   def destroy
     like = Like.find(params[:id])
-    if like.user.id == current_user.id
-      like.destroy
-      @prototype = Prototype.find(params[:prototype_id])
-    end
+    like.destroy if like.user.id == current_user.id
   end
 
   private
 
   def create_params
     params.permit(:prototype_id)
+  end
+
+  def set_prototype
+    @prototype = Prototype.find(params[:prototype_id])
   end
 
 end
